Added comments to GraphicLaneManager

+++ Factory/src/factory/graphics/GraphicLaneManager.java
@@ -30,32 +30,58 @@ public class GraphicLaneManager{
 	ArrayList <Boolean> lane2QueueTaken;			//The queue
 
 	//variables
-	int vX, vY;		//velocities
-	int lane_xPos, lane_yPos;	//lane relative position
-	boolean laneStart;			//default = false
-	boolean divergeUp;			//true - up, false - down
-	int timerCount;				//counter to periodic add item to lane
-	int binItemCount;			//current item in bin to dump
-	int vibrationCount;			//every 2 paint, it'll vibrate
+	/**The velocities of the Lane**/
+	int vX, vY;
+	/**The relative x and y position of the beginning of the LaneManager**/
+	int lane_xPos, lane_yPos;
+	/**boolean for when lane Start, default is false**/
+	boolean laneStart;
+	/**boolean for when feeder feeds to Top or bottom lane, true - up, false - down**/
+	boolean divergeUp;
+	/**Counter to periodic add item to lane**/
+	int timerCount;
+	/**Counter for vibration of items in lane**/
+	int vibrationCount;
+	/**The height of the vibration of items**/
 	private int vibrationAmplitude;
-	int laneManagerID;					//lane Manager Number
+	/**lane Manager Number**/
+	int laneManagerID;	
+	/**Counters to keep track of lane animation. 7 sprites**/
 	int laneAnimationCounter, laneAnimationSpeed;
+	/**Counter for diverger light animation**/
 	int divergerLightAnimationCounter;
+	/**boolean for the light animation. true is top. false is bottom.**/
 	boolean divergerLightAnimationCountUp;
-	boolean feederOn;			//Feeder on/off
-	boolean binExists;			
-	boolean lane1PurgeOn;
-	boolean lane2PurgeOn;
+	/**boolean for feeder is on. true - on. false - off**/
+	boolean feederOn;
+	/**Feeder has a bin**/
+	boolean binExists;		
+	/**boolean for Lane is being purged**/
+	boolean lane1PurgeOn,lane2PurgeOn;
+	/**boolean for feeder is being purged**/
 	boolean feederPurged;
+	/**counter for feeder to purge**/
 	int feederPurgeTimer;
+	/**Counter for stabilization**/
 	int stabilizationCount[];
+	/**boolean for when nest is stable. true - stable. false - not stable**/
 	boolean isStable[];
+	/**Lane speed for each lane. To be inputted from panel. Not for v1!**/
 	int laneSpeed;
+	/**The max distance the item must reach before changing direction**/
 	int itemYMax, itemXMax;
+	/**the coordinates of the expected location of items in lanes**/
 	int itemXLane, itemYLaneUp, itemYLaneDown;
-
+	/**Instance of the graphic panel**/
 	GraphicPanel graphicPanel;
 
+	/**
+	 * Create a Lane manager at the given x and y coordinates and ID
+	 * @param laneX The x coordinate of the lane manager
+	 * @param laneY the y coordinate of the lane manager
+	 * @param ID the ID number of the Lane Manager. 0-3
+	 * @param gp GraphicPanel for intercomponent communication
+	 */
 	public GraphicLaneManager(int laneX,int laneY, int ID, GraphicPanel gp){
 		lane_xPos = laneX; lane_yPos = laneY;					//MODIFY to change Lane position
 		laneManagerID = ID;
@@ -82,7 +108,7 @@ public class GraphicLaneManager{
 		lane1PurgeOn = false;		//Nest purge is off unless turned on
 		lane2PurgeOn = false;		//Nest purge is off unless turned on
 		feederPurged = false;
-		timerCount = 1; binItemCount = 0; vibrationCount = 0; vibrationAmplitude = 2;
+		timerCount = 1; vibrationCount = 0; vibrationAmplitude = 2;
 		stabilizationCount = new int[2];
 		isStable = new boolean[2];
 		for (int i = 0; i < 2; i++) {
@@ -103,6 +129,10 @@ public class GraphicLaneManager{
 		divergerLightOnImage = new ImageIcon("Images/divergerLights/dLight0.png");
 	}	
 
+	/**
+	 * Sets the Bin for the Lane manager's feeder
+	 * @param bin The Feeder's GraphicBin
+	 */
 	public void setBin(GraphicBin bin){
 		this.bin = bin;
 		if (bin != null) {
@@ -113,16 +143,27 @@ public class GraphicLaneManager{
 			feederPurged = false;
 		}
 	}
-
+	/**
+	 * Gets the Bin from the Lane Manager's feeder
+	 * @return the lane manager's feeder's bin
+	 */
 	public GraphicBin getBin(){
 		return bin;
 	}
 	
+	/**
+	 * returns the boolean for hasBin
+	 * @return return binExists
+	 */
 	public boolean hasBin()
 	{
 		return binExists;
 	}
 	
+	/**
+	 * Erase and return the current bin in the lane manager's feeder
+	 * @return the copy of the bin
+	 */
 	public GraphicBin popBin()
 	{
 		GraphicBin binCopy = bin;
@@ -131,6 +172,9 @@ public class GraphicLaneManager{
 		return binCopy;
 	}
 	
+	/**
+	 * Turns the purging of feeder on
+	 */
 	public void purgeFeeder() {
 		bin.getBinItems().clear();
 		feederOn = false;
@@ -138,6 +182,10 @@ public class GraphicLaneManager{
 		feederPurgeTimer = 0;
 	}
 	
+	/**
+	 * Paints the Lanes and the items within the lanes
+	 * @param g the specified graphics window
+	 */
 	public void paintLane(Graphics g){
 		Graphics2D g2 = (Graphics2D) g;
 
@@ -159,7 +207,11 @@ public class GraphicLaneManager{
 			lane2Items.get(i).paint(g2);
 		vibrationCount++;
 	} // END Paint function
-
+	
+	/**
+	 * paints the feeder, also alternates the lights of the diverger
+	 * @param g
+	 */
 	public void paintFeeder(Graphics g) {
 		// Draw background
 		g.setColor(Color.WHITE);
@@ -192,6 +244,10 @@ public class GraphicLaneManager{
 			divergerLightAnimationCounter -= 1;
 	}
 
+	/**
+	 * processes the lane manager.
+	 * Does everything for the Lane Manager
+	 */
 	public void moveLane() {
 		for (int i = 0; i < 2; i++) {
 			stabilizationCount[i]++;
@@ -213,7 +269,7 @@ public class GraphicLaneManager{
 				graphicPanel.purgeFeederDone(laneManagerID);
 			}
 		}
-		if(binExists){
+		if(binExists){	//If Bin exists, gets items from bin
 			if(laneStart){
 				if(feederOn){
 					if(timerCount % 10 == 0){		//Put an item on lane on a timed interval
@@ -247,16 +303,21 @@ public class GraphicLaneManager{
 			}
 			timerCount++;
 		}
-		else{
+		else{	//if bin does not exists, processes the items in the lanes
 			processLane();
 		}
 	}
 
+	/**
+	 * Moves the items in the lane.
+	 * Processes both top and bottom lane
+	 * DOES EVERYTHING
+	 */
 	public void processLane(){
 
 		if(lane1PurgeOn){		//If purge is on, empties the nest and destroys items on lane
 			graphicPanel.getNest().get(laneManagerID * 2).clearItems();
-			for(int j = 0; j <lane1Items.size();j++){
+			for(int j = 0; j <lane1Items.size();j++){	//changes the movement of the items for purging
 				if(lane1Items.get(j).getStepY() > 0){
 					lane1Items.get(j).setVY(-8);
 				}
@@ -264,12 +325,9 @@ public class GraphicLaneManager{
 					lane1Items.get(j).setVX(vX);
 				}
 			}
-			System.out.println("size " + lane1QueueTaken.size());
-			
-			if(lane1Items.size() == 0){
+			if(lane1Items.size() == 0){	//If lane is empty, purges nest and end 
 				lane1PurgeOn = false; //This is where the purge ends
 				lane1QueueTaken.clear();
-				System.out.println("size " + lane1QueueTaken.size());
 				graphicPanel.purgeTopLaneDone(laneManagerID);
 			}
 			else{
@@ -293,7 +351,7 @@ public class GraphicLaneManager{
 									lane1Items.get(i).setX(itemXLane);
 						}
 						lane1Items.get(i).setStepY(lane1Items.get(i).getStepY() - 1);
-						if(lane1Items.get(i).getStepY() == 0){
+						if(lane1Items.get(i).getStepY() == 0){	//Hits the end of Vertical movements. Turn and move horizontal
 							lane1Items.get(i).setVY(0);
 							lane1Items.get(i).setVX(vX);
 						}
@@ -317,12 +375,12 @@ public class GraphicLaneManager{
 								lane1Items.get(i).setY(itemYLaneUp);
 							}
 						}
-						if(lane1Items.get(i).getStepX() == 0){
+						if(lane1Items.get(i).getStepX() == 0){	//Purges the item. It disappears
 							lane1Items.remove(i);
 							i--;
 						}
 					}
-					if(lane1Items.size() == 0){
+					if(lane1Items.size() == 0){	//End of purging
 						lane1PurgeOn = false; //This is where the purge ends
 						lane1QueueTaken.clear();
 						graphicPanel.purgeTopLaneDone(laneManagerID);
@@ -360,7 +418,7 @@ public class GraphicLaneManager{
 				}
 				//Queue entering Nests
 				if(graphicPanel.getNest().get(laneManagerID * 2).getSize() < 9){
-					for(int j = 0; j <lane1Items.size();j++){
+					for(int j = 0; j <lane1Items.size();j++){	//changes the direction of item based on its location
 						if(lane1Items.get(j).getStepY() > 0){
 							lane1Items.get(j).setVY(-8);
 						}
@@ -368,7 +426,7 @@ public class GraphicLaneManager{
 							lane1Items.get(j).setVX(vX);
 						}
 					}
-					if(lane1Items.get(i).getStepX() == 0){
+					if(lane1Items.get(i).getStepX() == 0){	//enters the nest
 						lane1Items.get(i).setVY(0);
 						lane1Items.get(i).setVX(0);
 						stabilizationCount[0] = 0;
@@ -383,20 +441,21 @@ public class GraphicLaneManager{
 						i--;
 					}
 				}
-				else{
-					for(int j = 0; j <lane1Items.size();j++){
+				else{	//Nest is full
+					for(int j = 0; j <lane1Items.size();j++){	//sets item to pause when reaches the queue
 						if(lane1Items.get(j).getStepX() < lane1QueueTaken.size() + 1){
 							lane1Items.get(j).setVX(0);
 						}
 					}
-					if(lane1Items.get(i).getStepX() == lane1QueueTaken.size() + 1){		//IT IS THIS LINE TO FIX
+					if(lane1Items.get(i).getStepX() == lane1QueueTaken.size() + 1){		//if reaches queue
 						//Queue is full, delete crashing Items
-						if(lane1QueueTaken.size() > 17){ // To be changed according to size of lane
+						if(lane1QueueTaken.size() > 17){ // To be changed according to size of queue
+														 // If queue is full, delete the item
 							lane1Items.remove(i);
 							i--;
 							continue;
 						}
-						else{
+						else{	//adds item to queue, makes it stop moving horizontally.
 							lane1Items.get(i).setVX(0);
 							lane1QueueTaken.add(new Boolean(true));
 							
@@ -425,7 +484,7 @@ public class GraphicLaneManager{
 						}
 					}
 
-					if(graphicPanel.getNest().get(laneManagerID * 2).getSize() >= 9){
+					if(graphicPanel.getNest().get(laneManagerID * 2).getSize() >= 9){	//if lane is full
 						if(lane1Items.get(i).getStepX() == lane1QueueTaken.size() + 1){
 							//Queue is full, delete crashing Items
 							if(lane1QueueTaken.size() > 17){ // To be changed according to size of lane
@@ -439,7 +498,8 @@ public class GraphicLaneManager{
 							}
 						}
 					}
-					else if(lane1Items.get(i).getStepX() == 0){
+					else if(lane1Items.get(i).getStepX() == 0){	//lane is not full
+																//if item reaches nest, add to nest
 						lane1Items.get(i).setVY(0);
 						lane1Items.get(i).setVX(0);
 						stabilizationCount[0] = 0;
@@ -477,7 +537,8 @@ public class GraphicLaneManager{
 			}
 		}  // END OF LANE 1
 
-
+		//Repeat for lane 2 with different variables and speed
+		
 		if(lane2PurgeOn){		//If purge is on, empties the nest and destroys items on lane
 			graphicPanel.getNest().get(laneManagerID * 2 + 1).clearItems();
 			for(int j = 0; j <lane2Items.size();j++){
